Prevent embedded database shutdown
Wrap all embedded DataSource in a proxy that ignores .shutdown(). This
change is required due to the fact that SPR-9713 changes Spring
so that it considers 'shutdown' in the same way as 'destroy'.

Since the underlying HSQLDB session may still be part of a pool it
is imperative that global shutdown() does not occur.